Moved route test fixtures to 'conftest.py'
